+++ src/components/Answer/Answer.tsx
@@ -1,0 +1,55 @@
+// SPDX-FileCopyrightText: 2023 Félix Robles <felix@sequentech.io>
+//
+// SPDX-License-Identifier: AGPL-3.0-only
+import React from "react"
+import {useAppDispatch, useAppSelector} from "../../store/hooks"
+import {Candidate, stringToHtml, isUndefined} from "ui-essentials"
+import {IAnswer, IElectionDTO} from "sequent-core"
+import Image from "mui-image"
+import {
+    selectBallotSelectionVoteChoice,
+    setBallotSelectionVoteChoice,
+} from "../../store/ballotSelections/ballotSelectionsSlice"
+
+export interface IAnswerProps {
+    answer: IAnswer
+    questionIndex: number
+    election: IElectionDTO
+    hasCategory?: boolean
+}
+
+export const Answer: React.FC<IAnswerProps> = ({answer, questionIndex, election, hasCategory}) => {
+    const selectionState = useAppSelector(
+        selectBallotSelectionVoteChoice(election.id, questionIndex, answer.id)
+    )
+    const dispatch = useAppDispatch()
+    const imageUrl = answer.urls.find((url) => "Image URL" === url.title)?.url
+    const infoUrl = answer.urls.find((url) => "URL" === url.title)?.url
+
+    const isChecked = () => !isUndefined(selectionState) && selectionState.selected > -1
+    const setChecked = (value: boolean) =>
+        dispatch(
+            setBallotSelectionVoteChoice({
+                election,
+                questionIndex,
+                voteChoice: {
+                    id: answer.id,
+                    selected: value ? 0 : -1,
+                },
+            })
+        )
+
+    return (
+        <Candidate
+            title={answer.text}
+            description={stringToHtml(answer.details)}
+            isActive={true}
+            checked={isChecked()}
+            setChecked={setChecked}
+            url={infoUrl}
+            hasCategory={hasCategory}
+        >
+            {imageUrl ? <Image src={imageUrl} duration={100} /> : null}
+        </Candidate>
+    )
+}
